chore: Move PlanTier to shared crate
